Switch from u64 to i32 (#84)
This ensures that the overall codebase is more consistent.

+++ 1/assets/1.5-finished-code.rs
@@ -9,7 +9,7 @@ mod incrementer {
     #[ink(storage)]
     pub struct Incrementer {
         value: i32,
-        my_value: ink_storage::collections::HashMap<AccountId, u64>,
+        my_value: ink_storage::collections::HashMap<AccountId, i32>,
     }
 
     impl Incrementer {
@@ -40,11 +40,11 @@ mod incrementer {
         }
 
         #[ink(message)]
-        pub fn get_mine(&self) -> u64 {
+        pub fn get_mine(&self) -> i32 {
             self.my_value_or_zero(&self.env().caller())
         }
 
-        fn my_value_or_zero(&self, of: &AccountId) -> u64 {
+        fn my_value_or_zero(&self, of: &AccountId) -> i32 {
             *self.my_value.get(of).unwrap_or(&0)
         }
     }
